add unicode_string helper
