+++ src/plugins/helpers/unicode_string.h
@@ -102,42 +102,31 @@
  *                                                                         *
  ***************************************************************************/
 #pragma once
+#include <libdrakvuf/libdrakvuf.h>
+#include <string>
 
-#include <libhook/libhook.hpp>
-#include "plugins.h"
-
-/*
- * These 2 templates convert member-function-pointer to class type.
- * It is required to properly call member-function-pointer.
- * Read more in <libhook/libhook.hpp>.
- */
-template <typename T>
-struct class_type;
-
-template <typename T, typename R, typename... Args>
-struct class_type<R (T::*)(Args...)>
+struct unicode_string
 {
-    using type = T;
+    explicit unicode_string(drakvuf_t drakvuf, addr_t str_addr, vmi_pid_t pid = 0)
+        : drakvuf{drakvuf}, str_addr{str_addr}, pid{pid}, contents{drakvuf_read_unicode_va(drakvuf, str_addr, pid)}
+    {}
+
+    operator std::string() const
+    {
+        if (!content)
+            content = std::string(static_cast<const char*>(raw->contents));
+        return content;
+    }
+
+    ~unicode_string()
+    {
+        vmi_free_unicode_str(raw);
+    }
+
+private:
+    drakvuf_t drakvuf;
+    addr_t str_addr;
+    vmi_pid_t pid;
+    unicode_string_t* raw;
+    std::string content;
 };
-
-/**
- * This class is only needed for better backwards compatibility.
- * The new hooking interface prefers to use member-functions as callbacks
- * which provides `this`.
- */
-class PluginResult : public libhook::CallResult
-{
-public:
-    PluginResult()
-        : libhook::CallResult()
-    {};
-
-    class pluginex* plugin_ = nullptr;
-};
-
-template<typename Plugin>
-Plugin* GetTrapPlugin(const drakvuf_trap_info_t* info)
-{
-    static_assert(std::is_base_of_v<pluginex, Plugin>, "Plugin must derive from pluginex");
-    return dynamic_cast<Plugin*>(libhook::GetTrapParams<PluginResult>(info)->plugin_);
-}
