+++ src/plugins/output_format/common.h
@@ -109,7 +109,7 @@
 
 #include <libdrakvuf/libdrakvuf.h>
 #include <plugins/private.h>
-#include <plugins/type_traits_helpers.h>
+#include <plugins/helpers/type_traits.h>
 
 #include <algorithm>
 #include <functional>
