+++ src/plugins/output_format/csvfmt.h
@@ -107,7 +107,7 @@
 
 #include "common.h"
 
-#include "plugins/type_traits_helpers.h"
+#include "plugins/helpers/type_traits.h"
 
 namespace csvfmt
 {
