+++ src/plugins/output_format/deffmt.h
@@ -107,7 +107,7 @@
 
 #include "common.h"
 
-#include "type_traits_helpers.h"
+#include "plugins/helpers/type_traits.h"
 
 namespace deffmt
 {
